Update to use setuptools_scm for release version generation
To support openstack-infra we store the release version in git
tags and use setuptools_scm to extract that tag in the setup
file when generating sdist and wheel files.

This patch includes the use of the new xstatic-release tool
which handles the majority of management of metadata to support
the infra release process.

Change-Id: I14e0b5986d3433afd796ea44e66dfafd5003e732